+++ src/index.ts
@@ -1,7 +1,9 @@
 import camelCase from './case/camelCase';
+import capitalize from './case/capitalize';
 
 const StringFormatter = {
   camelCase,
+  capitalize,
 };
 
 export default StringFormatter;