+++ src/case/lowerCase.ts
@@ -1,0 +1,9 @@
+import isInvalid from '../util/isInvalid';
+
+const lowerCase = (str: string) => {
+  if (isInvalid(str)) return '';
+
+  return str.toLowerCase();
+};
+
+export default lowerCase;