+++ src/index.ts
@@ -1,9 +1,11 @@
 import camelCase from './case/camelCase';
 import capitalize from './case/capitalize';
+import lowerCase from './case/lowerCase';
 
 const StringFormatter = {
   camelCase,
   capitalize,
+  lowerCase,
 };
 
 export default StringFormatter;