+++ src/index.ts
@@ -1,11 +1,13 @@
 import camelCase from './case/camelCase';
 import capitalize from './case/capitalize';
 import lowerCase from './case/lowerCase';
+import upperCase from './case/upperCase';
 
 const StringFormatter = {
   camelCase,
   capitalize,
   lowerCase,
+  upperCase,
 };
 
 export default StringFormatter;