+++ src/index.ts
@@ -3,6 +3,7 @@ import capitalize from './case/capitalize';
 import lowerCase from './case/lowerCase';
 import upperCase from './case/upperCase';
 import titleCase from './case/titleCase';
+import formatPhone from './format/formatPhone';
 
 import count from './count/count';
 
@@ -13,6 +14,7 @@ const StringFormatter = {
   lowerCase,
   upperCase,
   titleCase,
+  formatPhone,
 };
 
 export default StringFormatter;