+++ src/util/isInvalid.ts
@@ -1,5 +1,5 @@
-const isInvalid = (str: string) => {
-  return str === '' || str === null || str === undefined;
+const isInvalid = (value: string | number) => {
+  return value === '' || value <= 0 || value === null || value === undefined;
 };
 
 export default isInvalid;