+++ CHANGELOG.md
@@ -11,6 +11,13 @@ O formato é baseado em [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - **Fixed/Corrigido** para qualquer correção de bug.
 - **Security/Segurança** em caso de vulnerabilidades.
 
+## [Unreleased] - 
+
+### Added
+
+- Implementação da função count
+- Implementação da função formatPhone
+
 ## [1.1.0] - 2022-07-17
 
 ### Added